### 2019-01-08 b3c5096
Update paper.bib
[skip ci]
-Corrected small error in the bibtex entry formatting
