+++ papers/v2_paper/paper.bib
@@ -43,7 +43,7 @@ publisher = "Annual Reviews",
 }
 
 @misc {heiber2018excimontec,
-author = {Michael C. Heiber}
+author = {Michael C. Heiber},
 title = {Excimontec},
 YEAR = {2018},
 howpublished = {\url{https://github.com/MikeHeiber/Excimontec}},
