### 2019-01-08 76c0197
Update paper.bib
[skip ci]
-Corrected small error in the bibtex entry formatting
