+++ papers/v2_paper/paper.bib
@@ -119,7 +119,7 @@ number = "SAND2009-6226",
 institution = "Sandia National Laboratories",
 }
 
-@misc{plimpton2009ssparks
+@misc{plimpton2009ssparks,
 author = "Steve Plimpton and Aidan Thompson and Alex Slepoy",
 title = "SSPARKS",
 year = "2009",
