### 2019-01-08 b394e8b
Update paper.bib
[skip ci]
-Corrected small errors in the bibtex entry formatting
