+++ papers/v2_paper/paper.bib
@@ -155,8 +155,8 @@ booktitle = "Radiation Effects in Solids",
 year = "2007",
 chapter = "1",
 pages = "1-23",
-series = "{NATO} Science Series"
+series = "{NATO} Science Series",
 volume = "235",
 publisher = "Springer",
-editor = "K. E. Sickafus and E. A. Kotomin"
+editor = "K. E. Sickafus and E. A. Kotomin",
 }
