### 2019-01-08 5cec35f
Update Paper Files
[skip ci]
-Corrected bibtex formatting errors
-Corrected citation reference errors and text formatting errors
-Corrected a typo
