+++ papers/v2_paper/paper.bib
@@ -31,7 +31,7 @@ doi = "10.1016/0021-9991(76)90041-3",
 publisher = "Elsevier",
 }
 
-@article {gillespie2007arcp,
+@article {gillespie2007arpc,
 author = "Daniel T. Gillespie",
 title = "Stochastic Simulation of Chemical Kinetics",
 year = "2007",
@@ -65,14 +65,14 @@ address = "",
 
 @misc{hoffmann2013kmos,
 author = "Max K. Hoffmann",
-title = "kmos",
+title = "{kmos}",
 year = "2013",
 howpublished = "\url{https://github.com/mhoffman/kmos}",
 }
 
 @article{hoffmann2014cpc,
 author = "Max K. Hoffmann and Sebastian Matera and Karsten Reuter",
-title = "kmos: A lattice kinetic {M}onte {C}arlo framework",
+title = "{kmos}: A lattice kinetic {M}onte {C}arlo framework",
 journal = "Comput. Phys. Commun.",
 year = "2014",
 volume = "185",
