+++ papers/v2_paper/paper.md
@@ -18,7 +18,7 @@ bibliography: paper.bib
 
 # Summary
 
-Kinetic Monte Carlo (KMC) simulations are a powerful method for investigating the dynamics on non-equilibrium systems[@voter2007chapter] and have been used to help solve problems in a wide variety of scientific domains, including atomic simulations of epitaxial growth,[@martin-bragado2018pms] vacancy diffusion and grain growth in solids,[@plimpton2009report] opto-electronic mechanisms in disordered organic electronic devices,[@heiber2019chapter] chemical and ionic diffusion and reactions for heterogeneous catalysis[@stamatakis2015jpcm] and electrochemical systems,[@turner2015chapter] complex chemical reaction networks,[@gillespie2007arpc] and predator-prey population dynamics.[@dobramysl2018jpa] To tackle this diverse range of problems there are numerous open-source software tools with different levels of quality, flexibility, documentation, testing, and support. Of these, ``SPPARKS``,[@plimpton2009report,@plimpton2009ssparks] ``kmos``,[@hoffmann2014cpc,@hoffmann2013kmos] and ``KMCLib``[@leetmaa2014cpc,@leetma2014kmclib] and  are several other examples of high-quality general KMC framework software tools for particle simulations. There are also numerous closed-source tools that have been developed by various research groups and companies around the world. However, the ``KMC_Lattice`` library is uniquely designed to be a lightweight and flexible object-oriented C\texttt{++} library that allows developers to more easily create custom KMC simulation software packages for use with high performance computing resources. With detailed API documentation using ``Doxygen``, rigorous testing using ``googletest``, and continuous integration testing using ``TravisCI``, ``KMC_Lattice`` is built to be a reliable and scalable package that can be used by a wide variety of other open-source KMC software tools in the future. Currently, I am using the ``KMC_Lattice`` library as the foundation for a new KMC software package called ``Excimontec``, which is designed to simulate disordered organic electronic devices.[@heiber2018excimontec]
[…]
