### 2019-01-08 c2dea4d
Update paper.bib
[skip ci]
-Added some missing information about several of the book chapters
