+++ papers/v2_paper/paper.bib
@@ -60,7 +60,7 @@ publisher = "Woodhead Publishing",
 year = "2019",
 chapter = "10",
 pages = "309-347",
-address = "",
+doi = "10.1016/B978-0-08-102284-9.00010-3",
 }
 
 @misc{hoffmann2013kmos,
@@ -143,6 +143,8 @@ title = "Kinetic {M}onte {C}arlo Simulation of Electrochemical Systems",
 editor = "Abby L. Parrill and Kenny B. Lipkowitz",
 booktitle = "Reviews in Computational Chemistry",
 publisher = "John Wiley & Sons",
+edition = "1",
+volume = "28",
 year = "2015",
 chapter = "5",
 pages = "175-204",
@@ -158,5 +160,7 @@ pages = "1-23",
 series = "{NATO} Science Series",
 volume = "235",
 publisher = "Springer",
-editor = "K. E. Sickafus and E. A. Kotomin",
+address = "Dordrecht",
+editor = "K. E. Sickafus and E. A. Kotomin and B. P. Uberuaga",
+doi = "10.1007/978-1-4020-5295-8_1",
 }
