### 2019-01-08 #1161 @MikeHeiber
I have made some minor corrections to the paper, but everything is now compiled correctly and ready for review.  The joss-edits branch contains the latest version of repository.
@katyhuff Here are some suggested reviewers that I found from the list:
@arghdos
@myousefi2016
@jeffhammond
@jwuttke
