### 2019-01-08 bf623e0
Update Changelog
[skip ci]
-Added notes about latest round of changes
