### 2019-01-08 #1161 @katyhuff
Thanks for suggesting reviewers @MikeHeiber.

Now, @MikeHeiber , let's ping a few fitting reviewers, with computational physics, materials science, and c++ expertise. We can start the review once two or three of them accept the request to review.

Dear reviewers,

@myousefi2016 as suggested: Thank you for volunteering, Mehrdad!
@jeffhammond as suggested: Thanks for your quick response. I appreciate your candor. No need to review this one.
@mdoucet Mathieu, are you able to and interested in taking on this review?
@jwuttke : Joachim,  are you able to and interested in taking on this review?
@jochym : Pawel, are you able to and interested in taking on this review?

- Title: KMC_Lattice v2.0: An Object-Oriented C++ Library for Custom Kinetic Monte Carlo Simulations
- Summary: A lightweight, object-oriented C++ library that contains a general framework for creating custom kinetic Monte Carlo simulation software tools. (e.g. for simulating organic semiconductors)
- Topic: Computational Physics for Materials Science
- Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01161/joss.01161/10.21105.joss.01161.pdf
- Submitting author: @MikeHeiber (Michael Heiber)
- Repository: https://github.com/MikeHeiber/KMC_Lattice
- Version: v2.0.0-rc.2
