### 2019-01-09 #1161 @arfon
> @whedon generate pdf from branch joss-edits

@MikeHeiber - this @whedon feature has been a little buggy in the past - did this work for you (compiling from a named branch)?
