### 2019-01-09 #1161 @MikeHeiber
@arfon Also, for my last submission a few months ago, you had to compile the paper locally to get it to work, but this time everything worked as expected on my end.
