### 2019-01-12 #1168 @whedon
[REVIEW]: KMC_Lattice v2.0: An Object-Oriented C++ Library for Custom Kinetic Monte Carlo Simulations
**Submitting author:** @MikeHeiber (<a href="http://orcid.org/0000-0002-1567-5663">Michael Heiber</a>)
**Repository:** <a href="https://github.com/MikeHeiber/KMC_Lattice" target ="_blank">https://github.com/MikeHeiber/KMC_Lattice</a>
**Version:** v2.0.1
