### 2019-01-12 #1168 @MikeHeiber
For the latest draft of the paper, please see the joss-edits branch.
