### 2019-01-18 MikeHeiber/KMC_Lattice#62 @MikeHeiber
Fix default logfile enabling issue
The problem is that the default value for enabling logging is not defined in the Parameters_Simulation class header file, so the behavior of running the simulation without setting this parameter is undefined.  Even if a derived simulation class does not enable logging, the compiler may interpret it as being enabled, which can cause errors.
