### 2019-01-18 b5df5ed
Default Parameters Bugfix
- Resolves #62
-Added default values for all parameters in Parameters_Simulation and Parameters_Lattice to avoid undefined behavior with uninitialized parameters
-Added tests that the default values for the Parameters_Simulation and Parameters_Lattice objects are invalid parameters
