### 2019-01-18 MikeHeiber/KMC_Lattice#62 @MikeHeiber
This bug is resolved now on the joss-edits branch and will be merged into master with the rest of the joss-edits corrections once the paper review process is completed.
