### 2019-01-18 #1168 @MikeHeiber
For full transparency, @myousefi2016 identified a bug ([Issue #62](https://github.com/MikeHeiber/KMC_Lattice/issues/62)) in KMC_Lattice through testing out the [KMC_Lattice_example](https://github.com/MikeHeiber/KMC_Lattice_example/) example program.  The issue is now resolved in the KMC_Lattice joss-edits branch and will be merged into master with all other corrections/improvements once the review phase is completed.
