### 2019-01-22 #1168 @mdoucet
I'm done with my review. The code is well organized and the examples are pretty good. Installing and running the code was easy. There's enough references to dig deeper and understand the problem. Overall everything looks good!
