### 2019-01-22 #1168 @katyhuff
Thanks @myousefi2016 and @mdoucet for your reviews !

@MikeHeiber thank you for a strong submission and for engaging actively in the review process! I have looked over the paper, double checked all the DOI links, and have conducted a high level review of the code itself. Everything looks ship-shape to me.
Once you have your code and paper all ready in the release version (I assume v2.0.0?), it'll be good to double check the paper, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. When you update this thread with the DOI of the archive, I'll move forward with accepting the submission! Until then, now is your moment for final touchups!
