### 2019-01-23 4700bcb
Documentation Typo Corrections
-Fixed several typos and spelling mistakes in the documentation source files
-Updated the Doxygen documentation files
