### 2019-01-23 4df5c75
README Update for v2.0 Release
[skip ci]
-Updated changelog to note the changes in this commit
-Updated changelog heading for the changes coming in v2.0.0
-Updated README by changing current status to note the release of a stable v2.0
-Updated README by modifying the acknowledgements for v2.0
