+++ README.md
@@ -1,4 +1,4 @@
-# KMC_Lattice
+# KMC_Lattice 
 
 ## General Information
 
