### 2019-01-23 7172127
Joss edits (#63)
Resolves #59 and #62 
-Updated CHANGELOG with notes about the most recent changes
-Made several previous changelog notes more concise
-Corrected errors in the paper bibtex entry formatting and added some missing reference info 
-Fixed incorrect copyright statement years in test.cpp and test_mpi.cpp
-Added line to .gitignore file to ignore the x64 build directory
-Added default values for all parameters in Parameters_Simulation and Parameters_Lattice to avoid undefined behavior with uninitialized parameters
-Added tests that the default values for the Parameters_Simulation and 
[…]
