+++ papers/v2_paper/paper.bib
@@ -31,7 +31,7 @@ doi = "10.1016/0021-9991(76)90041-3",
 publisher = "Elsevier",
 }
 
-@article {gillespie2007arcp,
+@article {gillespie2007arpc,
 author = "Daniel T. Gillespie",
 title = "Stochastic Simulation of Chemical Kinetics",
 year = "2007",
@@ -43,7 +43,7 @@ publisher = "Annual Reviews",
 }
 
 @misc {heiber2018excimontec,
-author = {Michael C. Heiber}
+author = {Michael C. Heiber},
 title = {Excimontec},
 YEAR = {2018},
 howpublished = {\url{https://github.com/MikeHeiber/Excimontec}},
@@ -60,19 +60,19 @@ publisher = "Woodhead Publishing",
 year = "2019",
 chapter = "10",
 pages = "309-347",
-address = "",
+doi = "10.1016/B978-0-08-102284-9.00010-3",
 }
 
 @misc{hoffmann2013kmos,
 author = "Max K. Hoffmann",
-title = "kmos",
+title = "{kmos}",
 year = "2013",
 howpublished = "\url{https://github.com/mhoffman/kmos}",
 }
 
 @article{hoffmann2014cpc,
 author = "Max K. Hoffmann and Sebastian Matera and Karsten Reuter",
-title = "kmos: A lattice kinetic {M}onte {C}arlo framework",
+title = "{kmos}: A lattice kinetic {M}onte {C}arlo framework",
 journal = "Comput. Phys. Commun.",
 year = "2014",
 volume = "185",
@@ -119,7 +119,7 @@ number = "SAND2009-6226",
 institution = "Sandia National Laboratories",
 }
 
-@misc{plimpton2009ssparks
+@misc{plimpton2009ssparks,
 author = "Steve Plimpton and Aidan Thompson and Alex Slepoy",
 title = "SSPARKS",
 year = "2009",
@@ -143,6 +143,8 @@ title = "Kinetic {M}onte {C}arlo Simulation of Electrochemical Systems",
 editor = "Abby L. Parrill and Kenny B. Lipkowitz",
 booktitle = "Reviews in Computational Chemistry",
 publisher = "John Wiley & Sons",
+edition = "1",
+volume = "28",
 year = "2015",
 chapter = "5",
 pages = "175-204",
@@ -155,8 +157,10 @@ booktitle = "Radiation Effects in Solids",
 year = "2007",
 chapter = "1",
 pages = "1-23",
-series = "{NATO} Science Series"
+series = "{NATO} Science Series",
 volume = "235",
 publisher = "Springer",
-editor = "K. E. Sickafus and E. A. Kotomin"
+address = "Dordrecht",
+editor = "K. E. Sickafus and E. A. Kotomin and B. P. Uberuaga",
+doi = "10.1007/978-1-4020-5295-8_1",
 }
