### 2019-01-23 MikeHeiber/KMC_Lattice#62 @MikeHeiber
Resolved this issue in PR #63 .
