### 2019-01-23 #1168 @MikeHeiber
Thank you @katyhuff @myousefi2016 and @mdoucet for reviewing my software tool and making this a smooth process.  I look forward to doing this again soon once my Excimontec KMC software tool is complete.

KMC_Lattice v2.0.0 is now released and the paper looks ready to go.  v2.0.0 is archived on Zenodo with the following DOI:
10.5281/zenodo.2547139
