### 2019-01-29 #1168 @MikeHeiber
@danielskatz Looking over the final proof, I found a small typo error in the manuscript.  In one sentence, the word "the" appears twice in a row.  How should I go about correcting this?
