### 2019-01-29 b11f4ca
JOSS Paper Fix
-Corrected typo in paper.md
-Updated version_string in Version.cpp
-Updated Doxygen documentation files
-Updated changelog
