### 2019-01-29 cd304c9
Update changelog
-Updated changelog to note the change to the Doxyfile
