### 2019-01-29 a5454c9
Updated Readme
-Updated README latest stable verison info to v2.0.1
-Updated Changelog to note the README change
