### 2019-01-29 d9d2af9
Fix joss paper (#64)
* JOSS Paper Fix
-Corrected typo in paper.md
-Updated README latest stable version info to v2.0.1
-Updated version_string in Version.cpp to v2.0.1
-Updated Doxygen documentation files for v2.0.1
-Updated CHANGELOG to note all changes in v2.0.1
