+++ papers/v2_paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 Kinetic Monte Carlo (KMC) simulations are a powerful method for investigating the dynamics of non-equilibrium systems [@voter2007chapter] and have been used to help solve problems in a wide variety of scientific domains, including atomic simulations of epitaxial growth [@martin-bragado2018pms], vacancy diffusion and grain growth in solids [@plimpton2009report], opto-electronic mechanisms in disordered organic electronic devices [@heiber2019chapter], chemical and ionic diffusion and reactions for heterogeneous catalysis [@stamatakis2015jpcm] and electrochemical systems [@turner2015chapter], complex chemical reaction networks [@gillespie2007arpc], and predator-prey population dynamics [@dobramysl2018jpa]. To tackle this diverse range of problems there are numerous open-source software tools with different levels of quality, flexibility, documentation, testing, and support. Of these, ``SPPARKS`` [@plimpton2009report,@plimpton2009ssparks], ``kmos`` [@hoffmann2014cpc,@hoffmann2013kmos], and ``KMCLib`` [@leetmaa2014cpc,@leetmaa2014kmclib] are several other examples of high-quality general KMC framework software tools for particle simulations. There are also numerous closed-source tools that have been developed by various research groups and companies around the world. However, the ``KMC_Lattice`` library is uniquely designed to be a lightweight and flexible object-oriented C\texttt{++} library that allows developers to more easily create custom KMC simulation software packages for use with high performance computing resources. With detailed API documentation using ``Doxygen``, rigorous testing using ``googletest``, and continuous integration testing using ``TravisCI``, ``KMC_Lattice`` is built to be a reliable and scalable package that can be used by a wide variety of other open-source KMC software tools in the future. Currently, I am using the ``KMC_Lattice`` library as the foundation for a new KMC software package called ``Excimontec``, which is designed to simulate disordered organic electronic devices [@heiber2018excimontec].
 
-The ``KMC_Lattice`` library contains a number of base (parent) classes that must be extended to create derived (child) classes to implement a user's simulation model of choice. Users must define the object entities in the simulation model and all of the events (transitions) possible for each object type. For each event, users must define the conditions under which each event occurs and can also define custom rate constant functions. Custom rate constant calculation functions can allow complex interactions between objects or between the objects and the simulation environment (lattice sites). ``KMC_Lattice`` v2.0 allows users to create simulation models on a cubic lattice and combines the the rejection-free Bortz, Kalos, and Lebowitz (BKL) algorithm [@bortz1975jcp] and several variations of Gillespie's first-reaction method [@gillespie1976jcp] to efficiently propagate the simulation. It is also designed to use MPI functionality to efficiently parallelize calculations for gathering statistics of the behavior of the stochastic simulation and contains a number of utility functions for gathering and analyzing the data generated by the simulation. The source code for all releases is archived with Zenodo.
[…]
