+++ CHANGELOG.md
@@ -6,7 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 --------------------------------------------------------------------------------------------------------------------------------
 
-## [v2.0.0] - 2019-01-22
+## [v2.0.1] - 2019-01-29
+
+### Changed
+- Version - Updated current version string to v2.0.1
+- Doxyfile - Updated documentation version to v2.0.1
+- docs - Updated Doxygen documentation for v2.0.1
+- README - Updated information about latest stable release to v2.0.1
+
+### Fixed
+- CHANGELOG - v2.0.0 release date
+- papers/v2_paper/paper.md - Fixed typo
+
+## [v2.0.0] - 2019-01-23
 
 ### Added
 - .gitignore - Ignore statement for the Visual Studio x64 build directory
