### 2019-01-29 #1168 @MikeHeiber
@katyhuff @danielskatz The paper is now corrected, and a new release has been issued.  Please change the release to v2.0.1 with the DOI: 10.5281/zenodo.2552685
